+++ noipy/__init__.py
@@ -3,14 +3,15 @@
 
 # noipy
 # Copyright (c) 2013 Pablo O Vieira (povieira)
-# See README.rst and epl-v10.html for details.
+# See README.rst and LICENSE for details.
 
 """
 A simple script to update IP on No-IP DNS service. 
 """
-__title__ = 'noipy'
-__version__ = '1.2.1'
-__author__ = 'Pablo O Vieira'
-__email__ = 'email@povieira.com'
-__license__ = 'Eclipse Public License (EPL) - v1.0'
 
+__title__ = "noipy"
+__version_info__ = ('1', '2', '2')
+__version__ = ".".join(__version_info__)
+__author__ = "Pablo O Vieira"
+__email__ = "email@povieira.com"
+__license__ = "Apache License, Version 2.0"