+++ noipy/dnsupdater.py
@@ -3,7 +3,7 @@
 
 # noipy.dnsupdater
 # Copyright (c) 2013 Pablo O Vieira (povieira)
-# See README.rst and epl-v10.html for details.
+# See README.rst and LICENSE for details.
 
 from __future__ import print_function
 
@@ -46,12 +46,10 @@ class DnsUpdaterPlugin(object):
         return self._hostname
 
     def _get_base_url(self):
-        """ (None) -> str
-
-        Get the base URL for DDNS Update API. URL must contain 'hostname'
+        """Get the base URL for DDNS Update API. URL must contain 'hostname'
         and 'ip'. If authentication is via token string 'token' argument must
-        be provided as well.
-        Example: https://{auth_str}@ddnsprovider.com/update?hostname={hostname}&ip={ip}
+        be provided as well. Example:
+          http://{auth_str}@ddnsprovider.com/update?hostname={hostname}&ip={ip}
 
         This method must be implemented by plugin subclasses
         """
@@ -59,9 +57,7 @@ class DnsUpdaterPlugin(object):
         return NotImplemented
 
     def update_dns(self, new_ip):
-        """(str) -> None
-
-        Call No-IP API based on dict login_info and return the status code.
+        """Call No-IP API based on dict login_info and return the status code.
         """
 
         if self.auth_type == 'T':
@@ -73,7 +69,8 @@ class DnsUpdaterPlugin(object):
             api_call_url = self._get_base_url().format(hostname=self.hostname,
                                                        ip=new_ip)
             request = urllib2.Request(api_call_url)
-            request.add_header('Authorization', 'Basic %s' % self.auth.base64key.decode('utf-8'))
+            request.add_header('Authorization', 'Basic %s' %
+                               self.auth.base64key.decode('utf-8'))
 
         try:
             response = urllib2.urlopen(request)
@@ -83,28 +80,35 @@ class DnsUpdaterPlugin(object):
 
     @property
     def status_message(self):
-        """(str) -> NoneType
+        """Return friendly response from API based on response code. """
 
-        Return friendly response from API based on response code.
-        """
-
-        msg = ''
+        msg = None
         if self.last_status_code in ['badauth', 'nochg', '401', '403']:
-            msg = "ERROR: Invalid username or password (%s)." % self.last_status_code
-        elif 'good' in self.last_status_code or 'nochg' in self.last_status_code:
+            msg = "ERROR: Invalid username or password (%s)." % \
+                  self.last_status_code
+        elif 'good' in self.last_status_code \
+                or 'nochg' in self.last_status_code:
             ip = re.search(r'(\d{1,3}\.?){4}', self.last_status_code).group()
             if 'good' in self.last_status_code:
                 msg = "SUCCESS: DNS hostname IP (%s) successfully updated." % ip
             else:
-                msg = "SUCCESS: IP address (%s) is up to date, nothing was changed. Additional 'nochg' updates may be considered abusive." % ip
+                msg = "SUCCESS: IP address (%s) is up to date, nothing was " \
+                      "changed. Additional 'nochg' updates may be considered" \
+                      " abusive." % ip
         elif self.last_status_code == '!donator':
-            msg = "ERROR: Update request include a feature that is not available to informed user."
+            msg = "ERROR: Update request include a feature that is not " \
+                  "available to informed user."
         elif self.last_status_code == 'notfqdn':
-            msg = "ERROR: The hostname specified is not a fully-qualified domain name (not in the form hostname.dyndns.org or domain.com)."
+            msg = "ERROR: The hostname specified is not a fully-qualified " \
+                  "domain name (not in the form hostname.dyndns.org or " \
+                  "domain.com)."
         elif self.last_status_code == 'nohost':
-            msg = "ERROR: Hostname specified does not exist in this user account."
+            msg = "ERROR: Hostname specified does not exist in this user " \
+                  "account."
         elif self.last_status_code == 'numhost':
-            msg = "ERROR: Too many hosts (more than 20) specified in an update. Also returned if trying to update a round robin (which is not allowed)."
+            msg = "ERROR: Too many hosts (more than 20) specified in an " \
+                  "update. Also returned if trying to update a round robin " \
+                  "(which is not allowed)."
         elif self.last_status_code == 'abuse':
             msg = "ERROR: Username/hostname is blocked due to update abuse."
         elif self.last_status_code == 'badagent':
@@ -112,7 +116,8 @@ class DnsUpdaterPlugin(object):
         elif self.last_status_code == 'dnserr':
             msg = "ERROR: DNS error encountered."
         elif self.last_status_code == '911':
-            msg = "ERROR: Problem on server side. Retry update in a few minutes."
+            msg = "ERROR: Problem on server side. Retry update in a few " \
+                  "minutes."
         elif self.last_status_code == 'OK':
             msg = "SUCCESS: DNS hostname successfully updated."
         elif self.last_status_code == 'KO':
@@ -127,28 +132,27 @@ class DnsUpdaterPlugin(object):
 
 
 class NoipDnsUpdater(DnsUpdaterPlugin):
-    """No-IP DDNS provider plugin
-    """
+    """No-IP DDNS provider plugin """
 
     auth_type = "P"
 
     def _get_base_url(self):
-        return "https://dynupdate.no-ip.com/nic/update?hostname={hostname}&myip={ip}"
+        return "https://dynupdate.no-ip.com/nic/update?hostname={hostname}" \
+               "&myip={ip}"
 
 
 class DynDnsUpdater(DnsUpdaterPlugin):
-    """DynDNS DDNS provider plugin
-    """
+    """DynDNS DDNS provider plugin """
 
     auth_type = "P"
 
     def _get_base_url(self):
-        return "http://members.dyndns.org/nic/update?hostname={hostname}&myip={ip}&wildcard=NOCHG&mx=NOCHG&backmx=NOCHG"
+        return "http://members.dyndns.org/nic/update?hostname={hostname}" \
+               "&myip={ip}&wildcard=NOCHG&mx=NOCHG&backmx=NOCHG"
 
 
 class DuckDnsUpdater(DnsUpdaterPlugin):
-    """DuckDNS DDNS provider plugin
-    """
+    """DuckDNS DDNS provider plugin """
 
     auth_type = "T"
 
@@ -161,4 +165,5 @@ class DuckDnsUpdater(DnsUpdaterPlugin):
         return hostname
 
     def _get_base_url(self):
-        return "https://www.duckdns.org/update?domains={hostname}&token={token}&ip={ip}"
+        return "https://www.duckdns.org/update?domains={hostname}" \
+               "&token={token}&ip={ip}"